fix(rust): parse numeric token directly (serde_json rounding gap); split test module
Two more items:

1. A real, empirically-verified parity bug: serde_json::Value::as_f64()
   rounds "9007199254740991.4" to 9007199254740992 (rejected by the
   safe-integer ceiling), while both V8's Number() and Rust's own std
   str::parse::<f64> round the identical literal to 9007199254740991
   (accepted) - serde_json's own number handling diverges from both at this
   exact boundary. Fixed by parsing the raw captured token text directly via
   str::parse::<f64> instead of routing through serde_json::Value, matching
   V8's rounding exactly. Added a regression test with the exact literal.

   This corrects an earlier verification mistake in this PR: the rounding
   check I ran previously tested Rust's std f64 parser directly, not the
   serde_json::Value::as_f64() path the classifier actually uses - a
   different, and in this specific case non-equivalent, code path.

2. Moved the 27 unit tests from an inline #[cfg(test)] module into
   tests/version_test.rs, matching this crate's own existing convention
   (fixtures_test.rs, lifecycle_test.rs test the public API from outside
   rather than an inline module) and bringing src/version.rs from 761 to
   462 lines, under the project's 700-line file-size guideline.
